+++ tests/Fixtures/RememberingMapper/RememberingMapper.php
@@ -14,7 +14,6 @@ declare(strict_types=1);
 namespace Rekalogika\Mapper\Tests\Fixtures\RememberingMapper;
 
 use Rekalogika\Mapper\Context\Context;
-use Rekalogika\Mapper\Exception\UnexpectedValueException;
 use Rekalogika\Mapper\MapperInterface;
 use Rekalogika\Mapper\ObjectCache\ObjectCacheFactoryInterface;
 use Rekalogika\Mapper\Transformer\Context\PresetMapping;
@@ -37,29 +36,20 @@ class RememberingMapper implements MapperInterface, ResetInterface
         $this->presetMapping = new PresetMapping();
     }
 
+    /** @psalm-suppress InvalidReturnType */
     public function map(object $source, object|string $target, ?Context $context = null): object
     {
         $objectCache = $this->objectCacheFactory->createObjectCache();
 
-        if ($context === null) {
-            $context = Context::create();
-        }
-
+        $context ??= Context::create();
         $context = $context->with($objectCache, $this->presetMapping);
 
         $result = $this->decorated->map($source, $target, $context);
 
-        if (is_object($target)) {
-            $target = $target::class;
-        }
-
-        if (!$result instanceof $target) {
-            throw new UnexpectedValueException(sprintf('Expected instance of "%s", got "%s"', $target, get_class($result)));
-        }
-
         $newPresetMapping = PresetMappingFactory::fromObjectCacheReversed($objectCache);
         $this->presetMapping->mergeFrom($newPresetMapping);
 
+        /** @psalm-suppress InvalidReturnStatement */
         return $result;
     }
 }