+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * perf(`ObjectToObjectTransformer`): Prevent delegating to `MainTransformer` if
   the current value in a dynamic property is a scalar.
 * feat(`PresetMappingFactory`): Add `fromObjectCache()` and `fromObjectCacheReversed()`.
+* chore: Simplify remembering mapper.
 
 ## 1.0.0